moving files around to match its purpose
