+++ python/src/coreai_models/models/ios/sam3/detr.py
@@ -22,8 +22,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from coreai_models.models.ios.sam3.primitives.layer_norm import LayerNormReauthored
-from coreai_models.models.ios.sam3.primitives.sdpa import BidirectionalSDPA
+from coreai_models.primitives.ios.bidirectional_sdpa import BidirectionalSDPA
+from coreai_models.primitives.ios.layer_norm import LayerNormReauthored
 
 _HIDDEN_SIZE = 256
 _NUM_HEADS = 8
