+++ python/src/coreai_models/primitives/ios/__init__.py
@@ -3,11 +3,14 @@
 # Use of this source code is governed by a BSD-3-clause license that can
 # be found in the LICENSE file or at https://opensource.org/licenses/BSD-3-Clause
 
+from coreai_models.primitives.ios.bidirectional_sdpa import BidirectionalSDPA
 from coreai_models.primitives.ios.cache import KVCacheHandler
 from coreai_models.primitives.ios.embedding import (
     GatherEmbeddings,
     LoadEmbeddings,
 )
+from coreai_models.primitives.ios.gelu import GELUReauthored, gelu_ane
+from coreai_models.primitives.ios.layer_norm import LayerNormReauthored
 from coreai_models.primitives.ios.mlp import MLP
 from coreai_models.primitives.ios.quantization import (
     dequantize_per_tensor,
@@ -18,13 +21,17 @@ from coreai_models.primitives.ios.rope import RoPECache
 from coreai_models.primitives.ios.sdpa import SDPA
 
 __all__ = [
-    "KVCacheHandler",
+    "BidirectionalSDPA",
+    "GELUReauthored",
     "GatherEmbeddings",
+    "KVCacheHandler",
+    "LayerNormReauthored",
     "LoadEmbeddings",
     "MLP",
     "RMSNorm",
     "RoPECache",
     "SDPA",
     "dequantize_per_tensor",
+    "gelu_ane",
     "quantize_per_tensor",
 ]
